+++ app/src/main/java/com/simple/launcher/retirement/utils/exts/SizeExt.kt
@@ -3,6 +3,7 @@ package com.simple.launcher.retirement.utils.exts
 import android.content.res.Resources
 import android.util.TypedValue
 
+
 private fun Float.spToPx(): Float {
 
     return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, this, Resources.getSystem().displayMetrics)
@@ -13,18 +14,20 @@ private fun Float.dpToPx(): Float {
     return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, this, Resources.getSystem().displayMetrics)
 }
 
+
 private val SP = HashMap<Float, Float>()
 
-fun Int.sp() = toFloat().sp()
+fun Int.sp() = toFloat().sp().toInt()
 
 fun Float.sp() = SP[this] ?: this.spToPx().apply {
 
     SP[this@sp] = this
 }
 
+
 private val DP = HashMap<Float, Float>()
 
-fun Int.dp() = toFloat().dp()
+fun Int.dp() = toFloat().dp().toInt()
 
 fun Float.dp() = DP[this] ?: this.dpToPx().apply {
